file-encrypter/requirements.txt
gostcrypto
pycryptodomex
cryptography
pyskein